+++ src/App.tsx
@@ -42,6 +42,7 @@ import Repetition from './screens/rehearsal/Repetition';
 import RepImport from './screens/rehearsal/RepImport';
 import RepConfig from './screens/rehearsal/RepConfig';
 import RepPlayer from './screens/rehearsal/RepPlayer';
+import RepCards from './screens/rehearsal/RepCards';
 import Scene from './screens/Scene';
 import Journal from './screens/Journal';
 import Carnet from './screens/Carnet';
@@ -94,6 +95,7 @@ export default function App() {
           <Route path="/repetition/nouveau" element={<RepImport />} />
           <Route path="/repetition/:id/config" element={<RepConfig />} />
           <Route path="/repetition/:id/jouer" element={<RepPlayer />} />
+          <Route path="/repetition/:id/cartes" element={<RepCards />} />
           <Route path="/scene" element={<Scene />} />
           <Route path="/journal" element={<Journal />} />
           <Route path="/carnet" element={<Carnet />} />